(fix): Move VersionResource to app.py and clean up comprehension
